Fixed Persistence Steps and Initial Negated Effects.
I also added some toString() to PlanGraphLiteral/Step for easier debugging. The issue I'm running into now is with Problem object. It seems that some Literals are being created twice (hence different instances of Literal which are not equal). For example the precondition of (Fly Rocket1 NOLA London) is (at Rocket1 Nola) which is instance of Predication. However, an effect of (Fly Rocket1 London Nola) is (at Rocket1 Nola) which is a different instance of Predication. For PlanGraph to work, only one instance of a certain predication must exist.