Minor Changes PlanGraph.toString().
Made some minor toString changes so that when we debug the mutex list, we get a better understanding of what's going on during each level. Got rid of an import from Operator causing a warning. Moved the creation of mutexlists into constructor in case we decide to make a plangraph that doesn't use mutex lists. Probably better to refactor and have two classes PlanGraph and PlanGraphMutex which extends PlanGraph (but that's later).